# server: fix tracing benchmark