--- a/pkg/server/node_tenant_test.go
+++ b/pkg/server/node_tenant_test.go
@@ -25,11 +25,11 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 )
 
-// TestMaybeRedactRecording verifies that redactRecording strips
-// sensitive details for recordings consumed by tenants.
+// Check that redactRecording strips sensitive details from recordings.
 //
-// See kvccl.TestTenantTracesAreRedacted for an end-to-end test of this.
-func TestRedactRecordingForTenant(t *testing.T) {
+// See kvccl.TestTenantTracesAreRedacted for an end-to-end test of tenant trace
+// redaction.
+func TestRedactRecording(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 
 	const (